backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
duckdb>=0.9.0
chromadb>=0.4.0
groq>=0.4.0
google-generativeai>=0.4.0
PyMuPDF>=1.23.0
pdfplumber>=0.10.0
pandas>=2.0.0
openpyxl>=3.1.0
Pillow>=10.0.0
pyyaml>=6.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
sentence-transformers>=2.2.0
